Token Deletion Code Review  (#1722)
## Description

Remaining code review change requests from
https://github.com/webstudio-is/webstudio-builder/pull/1695


## Code Review

- [ ] hi @TrySound , I need you to do
  - conceptual review (architecture, feature-correctness)
  - detailed review (read every line)
  - test it on preview

## Before requesting a review

- [ ] made a self-review
- [ ] added inline comments where things may be not obvious (the "why",
not "what")

## Before merging

- [ ] tested locally and on preview environment (preview dev login:
5de6)
- [ ] updated [test
cases](https://github.com/webstudio-is/webstudio-builder/blob/main/apps/builder/docs/test-cases.md)
document
- [ ] added tests
- [ ] if any new env variables are added, added them to `.env.example`
and the `builder/env-check.js` if mandatory

+++ apps/builder/app/builder/features/style-panel/style-source-section.tsx
@@ -379,9 +379,8 @@ export const StyleSourcesSection = () => {
           removeStyleSourceFromInstance(id);
         }}
         onDeleteItem={(id) => {
-          const token = availableStyleSources.find(
-            (source) => source.id === id
-          );
+          const styleSources = styleSourcesStore.get();
+          const token = styleSources.get(id);
           if (token?.type === "token") {
             setTokenToDelete(token);
           }
@@ -408,17 +407,17 @@ export const StyleSourcesSection = () => {
           renameStyleSource(item.id, item.label);
         }}
       />
-      {tokenToDelete && (
-        <DeleteConfirmationDialog
-          onClose={() => {
-            setTokenToDelete(undefined);
-          }}
-          onConfirm={() => {
+      <DeleteConfirmationDialog
+        token={tokenToDelete?.name}
+        onClose={() => {
+          setTokenToDelete(undefined);
+        }}
+        onConfirm={() => {
+          if (tokenToDelete) {
             deleteStyleSource(tokenToDelete.id);
-          }}
-          token={tokenToDelete.name}
-        />
-      )}
+          }
+        }}
+      />
     </>
   );
 };
@@ -426,7 +425,7 @@ export const StyleSourcesSection = () => {
 type DeleteConfirmationDialogProps = {
   onClose: () => void;
   onConfirm: () => void;
-  token: string;
+  token?: string;
 };
 
 const DeleteConfirmationDialog = ({
@@ -436,7 +435,7 @@ const DeleteConfirmationDialog = ({
 }: DeleteConfirmationDialogProps) => {
   return (
     <Dialog
-      open
+      open={token !== undefined}
       onOpenChange={(isOpen) => {
         if (isOpen === false) {
           onClose();
